fix data mapping

+++ src/main/java/com/free_open_university/backend/dao/UserDao.java
@@ -3,6 +3,5 @@ package com.free_open_university.backend.dao;
 import com.free_open_university.backend.bean.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDao extends JpaRepository<User, Long> {
+public interface UserDao extends JpaRepository<User, Integer> {
 }
-
