+++ src/main/java/com/free_open_university/backend/dao/UserHistoryDao.java
@@ -3,6 +3,5 @@ package com.free_open_university.backend.dao;
 import com.free_open_university.backend.bean.UserHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserHistoryDao extends JpaRepository<UserHistory, Long> {
+public interface UserHistoryDao extends JpaRepository<UserHistory, Integer> {
 }
-
